porting: making OpenDAL compilable on 32hf platforms
This patch should allow OpenDAL to compile on 32 bit platforms (hard float)
with minium side effects.

changes:

- opendal::types::capability::Capability:
    - write_multi_max_size
    - write_multi_min_size
    - write_multi_align_size
    Option<usize> -> Option<u64>

+++ core/src/types/capability.rs
@@ -87,15 +87,15 @@ pub struct Capability {
     /// write_multi_max_size is the max size that services support in write_multi.
     ///
     /// For example, AWS S3 supports 5GiB as max in write_multi.
-    pub write_multi_max_size: Option<usize>,
+    pub write_multi_max_size: Option<u64>,
     /// write_multi_min_size is the min size that services support in write_multi.
     ///
     /// For example, AWS S3 requires at least 5MiB in write_multi expect the last one.
-    pub write_multi_min_size: Option<usize>,
+    pub write_multi_min_size: Option<u64>,
     /// write_multi_align_size is the align size that services required in write_multi.
     ///
     /// For example, Google GCS requires align size to 256KiB in write_multi.
-    pub write_multi_align_size: Option<usize>,
+    pub write_multi_align_size: Option<u64>,
 
     /// If operator supports create dir.
     pub create_dir: bool,
